Rename to NetworkActionRow. Add test.
