--- ui/src/app/base/components/NetworkActions/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NetworkActions";
